Started working on player controller to implement turns

+++ Source/ConnectFour/CoinType.h
@@ -3,7 +3,7 @@
 #include "CoreMinimal.h"
 
 UENUM()
-enum class ECoinType : int32
+enum class ECoinType : uint8
 {
 	NONE,
 	YELLOW_COIN,
